+++ Infin8.Coapp.Dto/rptDividendPaid.cs
@@ -2,10 +2,10 @@
 {
     public  class rptDividendPaid
     {
-        public int Mem_Id { get; set; }
+        public decimal Mem_Id { get; set; }
         public string? MemberNo { get; set; }
         public string? MemberName { get; set; }
-        public int PbleMaster_Id { get; set; }
+        public decimal PbleMaster_Id { get; set; }
         public DateTime FromDate { get; set; }
         public DateTime ToDate { get; set; }
         public double DividendPaid { get; set; }
