+++ Infin8.Coapp.Dto/rptDividendTDFWDWorking.cs
@@ -2,9 +2,9 @@
 {
     public class rptDividendTDFWDWorking
     {
-        public int PbleMaster_Id { get; set; }
+        public decimal PbleMaster_Id { get; set; }
         public int PbleType { get; set; }
-        public int Mem_Id { get; set; }
+        public decimal Mem_Id { get; set; }
         public string? MemberNo { get; set; }
         public string? PerNo { get; set; }
         public string? MemberName { get; set; }
